Extract common inbox components that can be reused
