+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/inbox/InboxNoteActionEvent.kt
@@ -1,0 +1,7 @@
+package com.woocommerce.android.ui.inbox
+
+import com.woocommerce.android.viewmodel.MultiLiveEvent.Event
+
+sealed class InboxNoteActionEvent : Event() {
+    data class OpenUrlEvent(val url: String) : InboxNoteActionEvent()
+}
