+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/inbox/InboxViewModel.kt
@@ -1,6 +1,5 @@
 package com.woocommerce.android.ui.inbox
 
-import androidx.annotation.ColorRes
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.SavedStateHandle
@@ -18,10 +17,6 @@ import com.woocommerce.android.analytics.AnalyticsTracker.Companion.KEY_IS_LOADI
 import com.woocommerce.android.analytics.AnalyticsTracker.Companion.VALUE_INBOX_NOTE_ACTION_DISMISS
 import com.woocommerce.android.analytics.AnalyticsTracker.Companion.VALUE_INBOX_NOTE_ACTION_DISMISS_ALL
 import com.woocommerce.android.analytics.AnalyticsTracker.Companion.VALUE_INBOX_NOTE_ACTION_OPEN
-import com.woocommerce.android.ui.inbox.domain.InboxNote
-import com.woocommerce.android.ui.inbox.domain.InboxNote.NoteType.SURVEY
-import com.woocommerce.android.ui.inbox.domain.InboxNote.Status.ACTIONED
-import com.woocommerce.android.ui.inbox.domain.InboxNoteAction
 import com.woocommerce.android.ui.inbox.domain.InboxRepository
 import com.woocommerce.android.util.DateUtils
 import com.woocommerce.android.viewmodel.MultiLiveEvent.Event
@@ -31,8 +26,6 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
-import org.wordpress.android.util.DateTimeUtils
-import java.util.Date
 import javax.inject.Inject
 
 @HiltViewModel
@@ -124,7 +117,14 @@ class InboxViewModel @Inject constructor(
     private fun inboxNotesLocalUpdates() =
         inboxRepository.observeInboxNotes()
             .map { inboxNotes ->
-                val notes = inboxNotes.map { it.toInboxNoteUi() }
+                val notes = inboxNotes.map {
+                    it.toInboxNoteUi(
+                        dateUtils,
+                        resourceProvider,
+                        ::handleInboxNoteAction,
+                        ::dismissNote
+                    )
+                }
                 InboxState(
                     isLoading = false,
                     notes = notes,
@@ -132,59 +132,6 @@ class InboxViewModel @Inject constructor(
                     isRefreshing = false
                 )
             }
-
-    private fun InboxNote.toInboxNoteUi() =
-        InboxNoteUi(
-            id = id,
-            title = title,
-            description = description,
-            dateCreated = formatNoteCreationDate(dateCreated),
-            isSurvey = type == SURVEY,
-            isActioned = status == ACTIONED,
-            actions = mapInboxActionsToUi(),
-        )
-
-    private fun InboxNote.mapInboxActionsToUi(): List<InboxNoteActionUi> {
-        if (type == SURVEY && status == ACTIONED) {
-            return emptyList()
-        }
-
-        val noteActionsUi = actions
-            .map { it.toInboxActionUi(id) }
-            .toMutableList()
-
-        addDismissActionIfMissing(noteActionsUi)
-
-        return noteActionsUi
-    }
-
-    private fun InboxNote.addDismissActionIfMissing(noteActionsUi: MutableList<InboxNoteActionUi>) {
-        if (!actionsHaveDismiss(noteActionsUi)) {
-            noteActionsUi.add(
-                InboxNoteActionUi(
-                    id = 0,
-                    parentNoteId = id,
-                    label = DEFAULT_DISMISS_LABEL,
-                    textColor = R.color.color_surface_variant,
-                    url = "",
-                    onClick = { actionId, noteId -> dismissNote(actionId, noteId) }
-                )
-            )
-        }
-    }
-
-    private fun actionsHaveDismiss(noteActionsUi: List<InboxNoteActionUi>) =
-        noteActionsUi.any { it.label == DEFAULT_DISMISS_LABEL }
-
-    private fun InboxNoteAction.toInboxActionUi(parentNoteId: Long) =
-        InboxNoteActionUi(
-            id = id,
-            parentNoteId = parentNoteId,
-            label = label,
-            textColor = getActionTextColor(),
-            url = url,
-            onClick = ::handleInboxNoteAction
-        )
 
     private fun handleInboxNoteAction(actionId: Long, noteId: Long) {
         val clickedNote = inboxState.value?.notes?.firstOrNull { noteId == it.id }
@@ -218,25 +165,25 @@ class InboxViewModel @Inject constructor(
     private fun dismissNote(actionId: Long, noteId: Long) {
         trackInboxNoteActionClicked(VALUE_INBOX_NOTE_ACTION_DISMISS)
         viewModelScope.launch {
-            updateDismissingActionState(noteId, actionId, isDimissing = true)
+            updateDismissingActionState(noteId, actionId, isDismissing = true)
             inboxRepository
                 .dismissNote(noteId)
                 .fold(
                     onFailure = { showSyncError() },
                     onSuccess = {}
                 )
-            updateDismissingActionState(noteId, actionId, isDimissing = false)
+            updateDismissingActionState(noteId, actionId, isDismissing = false)
         }
     }
 
-    private fun updateDismissingActionState(noteId: Long, actionId: Long, isDimissing: Boolean) {
+    private fun updateDismissingActionState(noteId: Long, actionId: Long, isDismissing: Boolean) {
         _inboxState.value = _inboxState.value?.copy(
             notes = _inboxState.value?.notes?.map { note ->
                 if (note.id == noteId) {
                     val updatedActions =
                         note.actions.map { action ->
                             if (action.id == actionId) {
-                                action.copy(isDismissing = isDimissing)
+                                action.copy(isDismissing = isDismissing)
                             } else {
                                 action
                             }
@@ -253,39 +200,6 @@ class InboxViewModel @Inject constructor(
         triggerEvent(Event.ShowSnackbar(R.string.inbox_screen_sync_error))
     }
 
-    private fun InboxNoteAction.getActionTextColor() =
-        if (isPrimary) {
-            R.color.color_secondary
-        } else {
-            R.color.color_surface_variant
-        }
-
-    @SuppressWarnings("MagicNumber", "ReturnCount")
-    private fun formatNoteCreationDate(createdDate: String): String {
-        val creationDate = dateUtils.getDateFromIso8601String(createdDate)
-        val now = Date()
-
-        val minutes = DateTimeUtils.minutesBetween(now, creationDate)
-        when {
-            minutes < 1 -> return resourceProvider.getString(R.string.inbox_note_recency_now)
-            minutes < 60 -> return resourceProvider.getString(R.string.inbox_note_recency_minutes, minutes)
-        }
-        val hours = DateTimeUtils.hoursBetween(now, creationDate)
-        when {
-            hours == 1 -> return resourceProvider.getString(R.string.inbox_note_recency_one_hour)
-            hours < 24 -> return resourceProvider.getString(R.string.inbox_note_recency_hours, hours)
-        }
-        val days = DateTimeUtils.daysBetween(now, creationDate)
-        when {
-            days == 1 -> return resourceProvider.getString(R.string.inbox_note_recency_one_day)
-            days < 30 -> return resourceProvider.getString(R.string.inbox_note_recency_days, days)
-        }
-        return resourceProvider.getString(
-            R.string.inbox_note_recency_date_time,
-            dateUtils.toDisplayMMMddYYYYDate(creationDate?.time ?: 0) ?: ""
-        )
-    }
-
     data class InboxState(
         val isLoading: Boolean = false,
         val notes: List<InboxNoteUi> = emptyList(),
@@ -293,27 +207,4 @@ class InboxViewModel @Inject constructor(
         val isRefreshing: Boolean = false
     )
 
-    data class InboxNoteUi(
-        val id: Long,
-        val title: String,
-        val description: String,
-        val dateCreated: String,
-        val isSurvey: Boolean,
-        val isActioned: Boolean,
-        val actions: List<InboxNoteActionUi>
-    )
-
-    data class InboxNoteActionUi(
-        val id: Long,
-        val parentNoteId: Long,
-        val label: String,
-        @ColorRes val textColor: Int,
-        val url: String,
-        val isDismissing: Boolean = false,
-        val onClick: (Long, Long) -> Unit
-    )
-
-    sealed class InboxNoteActionEvent : Event() {
-        data class OpenUrlEvent(val url: String) : InboxNoteActionEvent()
-    }
 }
